+++ CHANGELOG.md
@@ -14,7 +14,10 @@ All user visible changes to `cucumber` crate will be documented in this file. Th
 ### Added
 
 - Ability for step functions to return `Result`. ([#151])
+- `writer::JUnit` behind the `output-junit` feature flag ([#147])
+- Arbitrary output for `writer::Basic` ([#147])
 
+[#147]: /../../pull/147
 [#151]: /../../pull/151